projects/AI-podcast/requirements.txt
# AI Podcast - Core dependencies
flask>=3.0.2
google-generativeai>=0.8.6
elevenlabs>=2.37.0
moviepy==1.0.3
pydub>=0.25.1
pytubefix>=10.3.8
requests>=2.31.0
python-dotenv>=1.0.1

# moviepy 1.0.3 compatible pins
decorator==4.4.2
imageio>=2.33.1
imageio-ffmpeg>=0.4.9
proglog>=0.1.10
numpy>=1.26.0
tqdm>=4.66.1
